2018-5-4
1.更换新车模,
2.增加两个垂直电感,用于圆环方向判断,但效果不是很好
3.修改转向模糊PID的Kp.Kd上限,转向更为平滑

+++ BSP/inc/bsp_motor.h
@@ -43,8 +43,8 @@ typedef struct
  	int16_t LeftPwm;				/*  左边电机PWM  */
 	int16_t RightPwm;				/*  右边电机PWM  */
 	
-	int16_t LeftEncoder;		/*  左边电机编码器  */
-	int16_t RightEncoder;		/*  右边电机编码器  */
+	int32_t LeftEncoder;		/*  左边电机编码器  */
+	int32_t RightEncoder;		/*  右边电机编码器  */
 	
 	float LeftSpeed;			/*  左边电机转速  */
 	float RightSpeed;			/*  右边电机转速  */
